order api and add product form control and  api impliment

+++ src/Components/Pages/Dashboard/ManageOrders/SingleOrder/SingleOrder.js
@@ -1,24 +1,36 @@
 import React from 'react';
+import Swal from 'sweetalert2';
 
-const SingleOrder = ({order}) => {
-    const {full_name,address,city,state,user_email,phone}=order;
-    const {title,price}=order.booking_deatils;
+const SingleOrder = ({ order }) => {
+    const { full_name, address, city, state, user_email, phone } = order;
+    const { title, price } = order.booking_deatils;
     const status = order.status.status;
 
     const handleDelete = id => {
-       
-           
-                const url = `https://arcane-ravine-56101.herokuapp.com/orders/${id}`
+        Swal.fire({
+            icon: "warning",
+            title: "Are you sure to delete this order?",
+            showCancelButton: true,
+            confirmButtonText: "Yes",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                const url = `https://young-shore-30046.herokuapp.com/orders/${id}`
                 fetch(url, {
                     method: 'DELETE'
                 })
+
                     .then(res => res.json())
                     .then(data => {
-                        // window.location.reload();
+                        if (data.deletedCount) {
+                            Swal.fire("Deleted!", "", "success");
+                            window.location.reload()
+                        }
                     })
-             
-      
-       
+            }
+        })
+
+
+
     }
 
 
@@ -27,7 +39,7 @@ const SingleOrder = ({order}) => {
         const update = {
             status: "Shipped"
         }
-        const url = `https://arcane-ravine-56101.herokuapp.com/ordersupdate/${id}`;
+        const url = `https://young-shore-30046.herokuapp.com/orders/${id}`;
         fetch(url, {
             method: 'PUT',
             headers: {
@@ -37,10 +49,12 @@ const SingleOrder = ({order}) => {
         })
             .then(res => res.json())
             .then(data => {
-                window.location.reload();
-
+                if (data.modifiedCount > 0) {
+                    Swal.fire("Updated!", "", "Success");
+                    window.location.reload()
+                }
             })
-        }
+    }
     let stutesbar = [];
     if (order.status.status === "Pending") {
         stutesbar = "Approve";
@@ -50,21 +64,21 @@ const SingleOrder = ({order}) => {
     }
 
     return (
-      <tbody>
-        <tr>
-            {/* <th scope="row">1</th> */}
-            <td>{full_name}</td>
-            <td>{title}</td>
-            <td>${price}</td>
-            <td>{user_email}</td>
-            <td>{address},{city},{state}</td>
-            <td>{phone}</td>
-            <td><button className="btn btn-primary "  onClick={()=> updateOrder(order._id)}>{stutesbar}</button></td>
-            <td><i onClick={()=> handleDelete(order._id)} className="fas fa-trash-alt delete-button text-center"></i></td>
-            
-                    
-        </tr>
-    </tbody>
+        <tbody>
+            <tr>
+                {/* <th scope="row">1</th> */}
+                <td>{full_name}</td>
+                <td>{title}</td>
+                <td>${price}</td>
+                <td>{user_email}</td>
+                <td>{address},{city},{state}</td>
+                <td>{phone}</td>
+                <td><button className="btn btn-primary " onClick={() => updateOrder(order._id)}>{stutesbar}</button></td>
+                <td><i onClick={() => handleDelete(order._id)} className="fas fa-trash-alt delete-button text-center"></i></td>
+
+
+            </tr>
+        </tbody>
     );
 };
